Branch out SplitTests
